Today I created some more entities/models for my project using the ERD (Entity Relationship Diagram) I had designed.
I have yet to add the relationships between entities and some of the bridging/junction entities I need. That will be my next step.

I have chosen to include the [key] data annotation above the primary key of each entity. This is to future-proof my code, as if I decide to rename my primary key to a naming convention that the framework doesn't automatically recognise as the primary key.