Worked on Relationships
Today, I worked on the relationships between the models in my project.

*Added relationship between category and item. (One to many)

*Added relationship between Order and Gift(One to One)

*Added relationship between item and order(many to many)

To find out why I chose to use these relationships please refer to my design in my documentation. I have provided a significant amount of detail about the design processes I followed and an clear explaintion of how my relationship between my models is setup.

I also added a 'bridge' model/entity for the 'order item' relationship to facilitate the many-to-many relationship between items and orders. This is necessary because multiple items can be associated with one order, and also, one item can be part of multiple orders.