Fixed bugs and made improvements.
Fixed the edit iItem controller view data list of categories as I had made a typo.

Add a description field to the create and edit controller as well as displayed the description on the details, edit, and create pages by modifying the html razor code.

I also added a "$" sign in front of prices in the sales data dashboard.

In my model before I had the field type “Decimal” when instead it should have been “decimal” with a lowercase “d”. I fixed this error.

I also ran the add migrations to create a new migration as I updated the field validation properties and data types in my database. I called this migration "ModelUpdates"

I also Made sure the description field that belongs to the item model/table name followed the proper naming conventions.