I added labels to the checkout form and worked the orders controllers.
I also decided it would be wise to move the shipping details to the orders table as a customer can have multiple gifts/orders and each order can have different shipping details.

I also worked on the I action result/method for the checkout, when the form is submitted by the user it will pass the inputs the user entered as paraments to the order controller checkout method which I created. The checkout method will then run the code I wrote to add the data to the database. The code consists of simple Linq queries that I have written to modify and add the data into the databaseficI added labels to the checkout form.

I also decided it would be wise to move the shipping details to the orders table as customers can have multiple gifts/orders and each order can have different shipping details.

I also worked on the action result/method for the checkout, when the form is submitted by the user it will pass the inputs the user entered as parameters to the order controller checkout method. The checkout method will then run the code I wrote to add the data to the database. The code consists of simple Linq queries that modify and add the data to the database.

I have not yet finished and plan on working on it soon.