Today, I made significant progress in developing the "Add to Cart" feature of my project. When the user clicks the button on the item's page, it passes the item ID to the newly created "Ordered Item Controller add to cart method I created.” There, my written code runs a LINQ statement to insert the product into the order item table (the cart).
However, there is still major work to be done regarding this feature. As of now, the users do not have their own cart, and the order item page displays all items ordered. Moreover, instead of increasing the quantity of an item that is already in the cart, it adds it twice. These are the bugs that I am yet to fix. I still need to link this to the user to the cart and pair it with my orders.