I created a return method that is called from the Add To Cart method. The method is called CheckUserOrders.
It will check the database to see if the signed-in user has any orders, and that the orders they have have the status id of 1 which is incomplete pending orders. The if statement checks if there are no orders that belong to the user and runs the LINQ query to create a new order with the user ID of the currently signed-in user. It will then return the string order ID back to the Add to Cart method where it is called from. However, if the user has a pending order (status 1) it will return the order ID of that order back to the Add to Cart method. This prevents the user from having more than one order until they have checkout as the items will keep getting added to the order until the order status changes when the user checks out.

P.S Today was the first day of the school holidays. I made significant progress in my project. There are obviously a lot of improvements which are coming soon. I am wanting to try to intergrate the stripe payment API into My project, I am not sure how diificault this process will be, however I am going to attempt it next.