Integrated stripe payment API into my project. This took a lot of time but reading through the stripe API documentation was helpful as it walked me through the steps of setting up the API. The frist thing I did was download the Stripe.Net package from the NuGet Package Manger.
I created a method called payment which is called from the checkout method checkout method which I now move from the order controller to the order item controller.

The payment Method communicates with the stripe API passing in information about the order needed for payment.

The documentation only provides instructions on how to pass the information of one product to the API. However in my project, often orders will have more than one product. Hence I coded a for loop which pass all the information about the products, their prices and the quantity ordered by the user. to the stripe payment API.

The way I achieved this was by creating a variable that runs a query and stores the order items that belong to the user Id passed in. To improve this in the future I will try and pass the ordered items in instead of having multiple queries in the order item controller doing the same thing.

 On the Stripe website, I had to create an account that would generate a private and public key and private key which I would then add to my project settings files. The stripe website also allows me to customise the colour of the payment page.

I have talked more about the stripe API Integration in my documentaion.

One issue I faced when integrating the stripe API into my project was that I added two constructors into one controller. One for the context and one for the stripe. However, this caused major issues as my program did not have which to pick from. After researching, I found out that having multiple constructors in the same class/controller is bad practice and will not work. To fix this I combined the two constructs into one. This fixed the errors and it was working.